feat: bind scanner to button
