+++ src/features/scan/components/ScanView.tsx
@@ -1,13 +1,50 @@
-import React, { FC } from 'react'
-import { View } from 'react-native'
+import React, { FC, useEffect, useRef } from 'react'
+import { Alert } from 'react-native'
+import {
+  Camera,
+  useCameraDevice,
+  useCameraPermission,
+  useCodeScanner,
+} from 'react-native-vision-camera'
 import styled from 'styled-components/native'
 
 import { TouchableOpacity } from 'ui/components/TouchableOpacity'
 
 export const ScanView: FC = () => {
+  const isScanned = useRef<boolean>(false)
+  const { hasPermission, requestPermission } = useCameraPermission()
+
+  const device = useCameraDevice('back')
+  const codeScanner = useCodeScanner({
+    codeTypes: ['qr', 'ean-13'],
+    onCodeScanned: (codes) => {
+      if (!isScanned.current) {
+        isScanned.current = true
+        Alert.alert(`Scanned ${codes.length} codes!`, `${codes[0]?.value}`, [
+          {
+            text: 'Fermer',
+            onPress: () => {
+              isScanned.current = false
+            },
+          },
+        ])
+      }
+    },
+  })
+
+  useEffect(() => {
+    if (!hasPermission) {
+      requestPermission()
+    }
+  }, [hasPermission, requestPermission])
+
   return (
     <Container>
-      <StyledCamera />
+      {hasPermission && device ? (
+        <StyledCamera device={device} isActive codeScanner={codeScanner} />
+      ) : (
+        <BlankScreen />
+      )}
       <ButtonContainer>
         <StyledButton>
           <ButtonText>Scan Text</ButtonText>
@@ -24,9 +61,17 @@ const Container = styled.View`
   flex: 1;
   justify-content: flex-end;
 `
-// met ton composant caméra ici
-// const StyledCamera = styled(RNCamera)`
-const StyledCamera = styled(View)`
+
+const StyledCamera = styled(Camera)`
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+`
+
+const BlankScreen = styled.View`
+  background-color: black;
   position: absolute;
   top: 0;
   left: 0;
